requirements.txt
Flask
Flask-DebugToolbar
Flask-SQLAlchemy
Flask-WTF
ipython
psycopg2-binary
python-dotenv
Flask-Bcrypt